requirements.txt
altgraph==0.17.4
auto-py-to-exe==2.46.0
bottle==0.13.4
bottle-websocket==0.2.9
cachetools==5.5.2
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
dotenv==0.9.9
Eel==0.18.2
future==1.0.0
gevent==25.5.1
gevent-websocket==0.10.1
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
greenlet==3.2.3
httplib2==0.22.0
idna==3.10
importlib_resources==6.5.2
oauthlib==3.3.1
packaging==25.0
pefile==2023.2.7
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.5
pyparsing==3.2.3
python-dotenv==1.1.1
pywin32-ctypes==0.2.3
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
setuptools==80.9.0
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
zope.event==5.1
zope.interface==7.2
